Break resolver tests up into several files.
