+++ pkg/graph/schema.resolvers_system_intake_test.go
@@ -1,0 +1,406 @@
+package graph
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/guregu/null"
+	_ "github.com/lib/pq" // required for postgres driver in sql
+
+	"github.com/cmsgov/easi-app/pkg/models"
+)
+
+func (s GraphQLTestSuite) TestFetchSystemIntakeQuery() {
+	ctx := context.Background()
+
+	// ToDo: This whole test would probably be better as an integration test in pkg/integration, so we can see the real
+	// functionality and not have to reinitialize all the services here
+	projectName := "Big Project"
+	businessOwner := "Firstname Lastname"
+	businessOwnerComponent := "OIT"
+
+	intake, intakeErr := s.store.CreateSystemIntake(ctx, &models.SystemIntake{
+		ProjectName:            null.StringFrom(projectName),
+		Status:                 models.SystemIntakeStatusINTAKESUBMITTED,
+		RequestType:            models.SystemIntakeRequestTypeNEW,
+		BusinessOwner:          null.StringFrom(businessOwner),
+		BusinessOwnerComponent: null.StringFrom(businessOwnerComponent),
+	})
+	s.NoError(intakeErr)
+
+	var resp struct {
+		SystemIntake struct {
+			ID            string
+			RequestName   string
+			Status        string
+			RequestType   string
+			BusinessOwner struct {
+				Name      string
+				Component string
+			}
+			BusinessOwnerComponent string
+			BusinessNeed           *string
+			BusinessCase           *string
+		}
+	}
+
+	// TODO we're supposed to be able to pass variables as additional arguments using client.Var()
+	// but it wasn't working for me.
+	s.client.MustPost(fmt.Sprintf(
+		`query {
+			systemIntake(id: "%s") {
+				id
+				requestName
+				status
+				requestType
+				businessOwner {
+					name
+					component
+				}
+				businessNeed
+			}
+		}`, intake.ID), &resp)
+
+	s.Equal(intake.ID.String(), resp.SystemIntake.ID)
+	s.Equal(projectName, resp.SystemIntake.RequestName)
+	s.Equal("INTAKE_SUBMITTED", resp.SystemIntake.Status)
+	s.Equal("NEW", resp.SystemIntake.RequestType)
+	s.Equal(businessOwner, resp.SystemIntake.BusinessOwner.Name)
+	s.Equal(businessOwnerComponent, resp.SystemIntake.BusinessOwner.Component)
+	s.Nil(resp.SystemIntake.BusinessNeed)
+	s.Nil(resp.SystemIntake.BusinessCase)
+}
+
+func (s GraphQLTestSuite) TestFetchSystemIntakeWithNotesQuery() {
+	ctx := context.Background()
+
+	// ToDo: This whole test would probably be better as an integration test in pkg/integration, so we can see the real
+	// functionality and not have to reinitialize all the services here
+	projectName := "Big Project"
+	businessOwner := "Firstname Lastname"
+	businessOwnerComponent := "OIT"
+
+	intake, intakeErr := s.store.CreateSystemIntake(ctx, &models.SystemIntake{
+		ProjectName:            null.StringFrom(projectName),
+		Status:                 models.SystemIntakeStatusINTAKESUBMITTED,
+		RequestType:            models.SystemIntakeRequestTypeNEW,
+		BusinessOwner:          null.StringFrom(businessOwner),
+		BusinessOwnerComponent: null.StringFrom(businessOwnerComponent),
+	})
+	s.NoError(intakeErr)
+
+	_, noteErr := s.store.CreateNote(ctx, &models.Note{
+		SystemIntakeID: intake.ID,
+		AuthorEUAID:    "QQQQ",
+		AuthorName:     null.StringFrom("Author Name"),
+		Content:        null.StringFrom("a clever remark"),
+	})
+	s.NoError(noteErr)
+
+	var resp struct {
+		SystemIntake struct {
+			ID    string
+			Notes []struct {
+				ID     string
+				Author struct {
+					Name string
+					EUA  string
+				}
+				Content   string
+				CreatedAt string
+			}
+		}
+	}
+
+	// TODO we're supposed to be able to pass variables as additional arguments using client.Var()
+	// but it wasn't working for me.
+	s.client.MustPost(fmt.Sprintf(
+		`query {
+			systemIntake(id: "%s") {
+				id
+				notes {
+					id
+					createdAt
+					content
+					author {
+						name
+						eua
+					}
+				}
+			}
+		}`, intake.ID), &resp)
+
+	s.Equal(intake.ID.String(), resp.SystemIntake.ID)
+
+	s.Len(resp.SystemIntake.Notes, 1)
+
+	s.NotEmpty(resp.SystemIntake.Notes[0].ID)
+	s.Equal("a clever remark", resp.SystemIntake.Notes[0].Content)
+	s.Equal("QQQQ", resp.SystemIntake.Notes[0].Author.EUA)
+	s.Equal("Author Name", resp.SystemIntake.Notes[0].Author.Name)
+	s.NotEmpty(resp.SystemIntake.Notes[0].CreatedAt)
+}
+
+func (s GraphQLTestSuite) TestFetchSystemIntakeWithContractMonthAndYearQuery() {
+	ctx := context.Background()
+
+	// ToDo: This whole test would probably be better as an integration test in pkg/integration, so we can see the real
+	// functionality and not have to reinitialize all the services here
+	contracStartMonth := "10"
+	contractStartYear := "2002"
+	contractEndMonth := "08"
+	contractEndYear := "2020"
+	projectName := "My cool project"
+
+	intake, intakeErr := s.store.CreateSystemIntake(ctx, &models.SystemIntake{
+		ProjectName:        null.StringFrom(projectName),
+		Status:             models.SystemIntakeStatusINTAKESUBMITTED,
+		RequestType:        models.SystemIntakeRequestTypeNEW,
+		ContractStartMonth: null.StringFrom(contracStartMonth),
+		ContractStartYear:  null.StringFrom(contractStartYear),
+		ContractEndMonth:   null.StringFrom(contractEndMonth),
+		ContractEndYear:    null.StringFrom(contractEndYear),
+	})
+	s.NoError(intakeErr)
+
+	var resp struct {
+		SystemIntake struct {
+			ID       string
+			Contract struct {
+				EndDate struct {
+					Day   string
+					Month string
+					Year  string
+				}
+				StartDate struct {
+					Day   string
+					Month string
+					Year  string
+				}
+			}
+		}
+	}
+
+	// TODO we're supposed to be able to pass variables as additional arguments using client.Var()
+	// but it wasn't working for me.
+	s.client.MustPost(fmt.Sprintf(
+		`query {
+			systemIntake(id: "%s") {
+				id
+				contract {
+					endDate {
+						day
+						month
+						year
+					}
+					startDate {
+						day
+						month
+						year
+					}
+				}
+			}
+		}`, intake.ID), &resp)
+
+	s.Equal(intake.ID.String(), resp.SystemIntake.ID)
+
+	startDate := resp.SystemIntake.Contract.StartDate
+	s.Equal("", startDate.Day)
+	s.Equal(contracStartMonth, startDate.Month)
+	s.Equal(contractStartYear, startDate.Year)
+
+	endDate := resp.SystemIntake.Contract.EndDate
+	s.Equal("", endDate.Day)
+	s.Equal(contractEndMonth, endDate.Month)
+	s.Equal(contractEndYear, endDate.Year)
+}
+
+func (s GraphQLTestSuite) TestFetchSystemIntakeWithContractDatesQuery() {
+	ctx := context.Background()
+
+	// ToDo: This whole test would probably be better as an integration test in pkg/integration, so we can see the real
+	// functionality and not have to reinitialize all the services here
+	projectName := "My cool project"
+	contractStartDate, _ := time.Parse("2006-1-2", "2002-8-24")
+	contractEndDate, _ := time.Parse("2006-1-2", "2020-10-31")
+
+	intake, intakeErr := s.store.CreateSystemIntake(ctx, &models.SystemIntake{
+		ProjectName:       null.StringFrom(projectName),
+		Status:            models.SystemIntakeStatusINTAKESUBMITTED,
+		RequestType:       models.SystemIntakeRequestTypeNEW,
+		ContractStartDate: &contractStartDate,
+		ContractEndDate:   &contractEndDate,
+	})
+	s.NoError(intakeErr)
+
+	var resp struct {
+		SystemIntake struct {
+			ID       string
+			Contract struct {
+				EndDate struct {
+					Day   string
+					Month string
+					Year  string
+				}
+				StartDate struct {
+					Day   string
+					Month string
+					Year  string
+				}
+			}
+		}
+	}
+
+	// TODO we're supposed to be able to pass variables as additional arguments using client.Var()
+	// but it wasn't working for me.
+	s.client.MustPost(fmt.Sprintf(
+		`query {
+			systemIntake(id: "%s") {
+				id
+				contract {
+					endDate {
+						day
+						month
+						year
+					}
+					startDate {
+						day
+						month
+						year
+					}
+				}
+			}
+		}`, intake.ID), &resp)
+
+	s.Equal(intake.ID.String(), resp.SystemIntake.ID)
+
+	startDate := resp.SystemIntake.Contract.StartDate
+	s.Equal("24", startDate.Day)
+	s.Equal("8", startDate.Month)
+	s.Equal("2002", startDate.Year)
+
+	endDate := resp.SystemIntake.Contract.EndDate
+	s.Equal("31", endDate.Day)
+	s.Equal("10", endDate.Month)
+	s.Equal("2020", endDate.Year)
+}
+
+func (s GraphQLTestSuite) TestFetchSystemIntakeWithNoCollaboratorsQuery() {
+	ctx := context.Background()
+
+	// ToDo: This whole test would probably be better as an integration test in pkg/integration, so we can see the real
+	// functionality and not have to reinitialize all the services here
+	projectName := "My cool project"
+
+	intake, intakeErr := s.store.CreateSystemIntake(ctx, &models.SystemIntake{
+		ProjectName:                 null.StringFrom(projectName),
+		Status:                      models.SystemIntakeStatusINTAKESUBMITTED,
+		RequestType:                 models.SystemIntakeRequestTypeNEW,
+		EACollaboratorName:          null.StringFrom(""),
+		OITSecurityCollaboratorName: null.StringFrom(""),
+		TRBCollaboratorName:         null.StringFrom(""),
+	})
+	s.NoError(intakeErr)
+
+	var resp struct {
+		SystemIntake struct {
+			ID              string
+			GovernanceTeams struct {
+				IsPresent bool
+				Teams     []struct {
+					Acronym      string
+					Collaborator string
+					Key          string
+					Label        string
+					Name         string
+				}
+			}
+		}
+	}
+
+	// TODO we're supposed to be able to pass variables as additional arguments using client.Var()
+	// but it wasn't working for me.
+	s.client.MustPost(fmt.Sprintf(
+		`query {
+			systemIntake(id: "%s") {
+				id
+				governanceTeams {
+					isPresent
+					teams {
+						acronym
+						collaborator
+						key
+						label
+						name
+					}
+				}
+			}
+		}`, intake.ID), &resp)
+
+	s.Equal(intake.ID.String(), resp.SystemIntake.ID)
+	s.False(resp.SystemIntake.GovernanceTeams.IsPresent)
+	s.Nil(resp.SystemIntake.GovernanceTeams.Teams)
+}
+
+func (s GraphQLTestSuite) TestFetchSystemIntakeWithCollaboratorsQuery() {
+	ctx := context.Background()
+
+	// ToDo: This whole test would probably be better as an integration test in pkg/integration, so we can see the real
+	// functionality and not have to reinitialize all the services here
+	projectName := "My cool project"
+	eaName := "My EA Rep"
+	oitName := "My OIT Rep"
+	trbName := "My TRB Rep"
+
+	intake, intakeErr := s.store.CreateSystemIntake(ctx, &models.SystemIntake{
+		ProjectName:                 null.StringFrom(projectName),
+		Status:                      models.SystemIntakeStatusINTAKESUBMITTED,
+		RequestType:                 models.SystemIntakeRequestTypeNEW,
+		EACollaboratorName:          null.StringFrom(eaName),
+		OITSecurityCollaboratorName: null.StringFrom(oitName),
+		TRBCollaboratorName:         null.StringFrom(trbName),
+	})
+	s.NoError(intakeErr)
+
+	var resp struct {
+		SystemIntake struct {
+			ID              string
+			GovernanceTeams struct {
+				IsPresent bool
+				Teams     []struct {
+					Acronym      string
+					Collaborator string
+					Key          string
+					Label        string
+					Name         string
+				}
+			}
+		}
+	}
+
+	// TODO we're supposed to be able to pass variables as additional arguments using client.Var()
+	// but it wasn't working for me.
+	s.client.MustPost(fmt.Sprintf(
+		`query {
+			systemIntake(id: "%s") {
+				id
+				governanceTeams {
+					isPresent
+					teams {
+						acronym
+						collaborator
+						key
+						label
+						name
+					}
+				}
+			}
+		}`, intake.ID), &resp)
+
+	s.Equal(intake.ID.String(), resp.SystemIntake.ID)
+	s.True(resp.SystemIntake.GovernanceTeams.IsPresent)
+	s.Equal(eaName, resp.SystemIntake.GovernanceTeams.Teams[0].Collaborator)
+	s.Equal(oitName, resp.SystemIntake.GovernanceTeams.Teams[1].Collaborator)
+	s.Equal(trbName, resp.SystemIntake.GovernanceTeams.Teams[2].Collaborator)
+}
